refactor values off seekbar to string

+++ features/audioeffects/src/bass/java/com/barryzeha/audioeffects/common/EqualizerManager.kt
@@ -42,13 +42,13 @@ object EqualizerManager {
             val bandValue = if (eqValue.toInt() != 30) eqValue else getEqualizerBandPreConfig(
                 prefs.effectType,
                 seekId
-            )
-            updateFX(seekId, bandValue.toInt())
+            ).toFloat()
+            updateFX(seekId, bandValue)
         }
         val reverbValue = prefs.getReverbSeekBandValue(prefs.effectType, coreRes.id.reverb)
-        updateFX(fxArray.size - 1, reverbValue.toInt())
+        updateFX(fxArray.size - 1, reverbValue)
         val volumeValue = prefs.getVolumeSeekBandValue(prefs.effectType, coreRes.id.volume)
-        updateFX(11, volumeValue.toInt())
+        updateFX(11, volumeValue)
     }
 
     fun setupFX(fxIndex:(index:Int)->Unit) {
@@ -67,25 +67,25 @@ object EqualizerManager {
             fxIndex(i)
         }
         fxArray[fxArray.size - 1] = BASS.BASS_ChannelSetFX(chan, BASS.BASS_FX_DX8_REVERB, 0)
-        updateFX(fxArray.size - 1, fxArray[fxArray.size - 1])
+        updateFX(fxArray.size - 1, fxArray[fxArray.size - 1].toFloat())
         mPrefs?.let {prefs->
             val volumeValue = prefs.getVolumeSeekBandValue(prefs.effectType, coreRes.id.volume)
             BASS.BASS_ChannelSetAttribute(chan, BASS.BASS_ATTRIB_VOL, volumeValue / 15f)
         }
     }
 
-    fun updateFX(index: Int, value: Int) {
+    fun updateFX(index: Int, value: Float) {
         val n = index
-        val v = value.toInt()
+        val v = value
         if (n < fxArray.size - 1) { // EQ
             val p: BASS.BASS_DX8_PARAMEQ = BASS.BASS_DX8_PARAMEQ()
             BASS.BASS_FXGetParameters(fxArray[n], p)
-            p.fGain = (v - 15).toFloat()
+            p.fGain = (v - 15f).toFloat()
             BASS.BASS_FXSetParameters(fxArray[n], p)
         } else if (n == fxArray.size - 1) { // reverb
             val p: BASS.BASS_DX8_REVERB = BASS.BASS_DX8_REVERB()
             BASS.BASS_FXGetParameters(fxArray[n], p)
-            p.fReverbMix = (if (v != 0) (Math.log(v / 30.0) * 30).toFloat() else (-96).toFloat())
+            p.fReverbMix = (if (v.toInt() != 0) (Math.log(v / 30.0) * 30).toFloat() else (-96).toFloat())
             BASS.BASS_FXSetParameters(fxArray[n], p)
         } else // volume
             BASS.BASS_ChannelSetAttribute(chan, BASS.BASS_ATTRIB_VOL, v / 15f)
